Improve decision about reparenting to /
One condition where we a file has an invisible parent id is when:
1) someone creates a file in a folder
2) they share that file with you
3) you add the file to your drive, optionally in some sub-folder

The file has two parents, one of which is visible to you in drive, and
the other is a docid that's invisible to you.  There is no need to make
that file accessible at the root just because one of its parents can't
be found, only if *all* of its parents can't be found.