Rip out the concept of 'Handles', return inode
'handleID's returned to fuse merely have to uniquely identify the file
described by Open, so they can uniqely describe the file again at
Read().  The cheapest way to ensure that is just return the Inode.
There seems to be no need to preserve the knowledge of which files the
kernel has asked about in the past.

This might have to come back when we implement Write...